pyproject.toml
[project]
name = "freshroles"
version = "0.1.0"
description = "Job posting discovery and tracking tool for multiple ATS systems"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "selectolax>=0.3.0",
    "beautifulsoup4>=4.12.0",
    "simhash>=2.1.0",
    "pyyaml>=6.0",
    "tenacity>=8.2.0",
    "playwright>=1.40.0",
]

[project.optional-dependencies]
embeddings = [
    "openai>=1.0.0",
    "numpy>=1.24.0",
]
local-embeddings = [
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    "numpy>=1.24.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-httpx>=0.28.0",
    "ruff>=0.1.0",
]
resume = [
    "pypdf>=4.0.0",
    "python-docx>=1.0.0",
]

[project.scripts]
freshroles = "freshroles.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
